Renamed dashboard to search and changed imports to reflect this
